feat!: remove activeDialect unused variable, move type-conversion
This PR's 'feat!' is synthetic, i use it for the major version upgrade release
