+++ src/autocomplete/lib/sql-reference/type-conversion.ts
